+++ lib/rubocop/cop/rails/reflection_class_name.rb
@@ -13,7 +13,9 @@ module RuboCop
       #
       #   # good
       #   has_many :accounts, class_name: 'Account'
-      class ReflectionClassName < Cop
+      class ReflectionClassName < Base
+        extend AutoCorrector
+
         MSG = 'Use a string value for `class_name`.'
         RESTRICT_ON_SEND = %i[has_many has_one belongs_to].freeze
 
@@ -24,13 +26,23 @@ module RuboCop
         PATTERN
 
         def_node_matcher :reflection_class_name, <<~PATTERN
-          (pair (sym :class_name) [!dstr !str !sym])
+          (pair (sym :class_name) $[!dstr !str !sym])
         PATTERN
 
         def on_send(node)
-          association_with_reflection(node) do |reflection_class_name|
-            add_offense(node, location: reflection_class_name.loc.expression)
+          association_with_reflection(node) do |reflection|
+            add_offense(reflection) do |corrector|
+              class_node = reflection_class_name(reflection)
+              corrector.replace(class_node, replacement(class_node))
+            end
           end
+        end
+
+        private
+
+        def replacement(class_node)
+          replacement_node = class_node.send_type? ? class_node.children.first : class_node
+          replacement_node.source.inspect
         end
       end
     end