+++ spec/rubocop/cop/rails/reflection_class_name_spec.rb
@@ -11,12 +11,20 @@ RSpec.describe RuboCop::Cop::Rails::ReflectionClassName do
         has_many :accounts, class_name: Account, foreign_key: :account_id
                             ^^^^^^^^^^^^^^^^^^^ Use a string value for `class_name`.
       RUBY
+
+      expect_correction(<<~RUBY)
+        has_many :accounts, class_name: "Account", foreign_key: :account_id
+      RUBY
     end
 
     it '.name' do
       expect_offense(<<~RUBY)
         has_many :accounts, class_name: Account.name
                             ^^^^^^^^^^^^^^^^^^^^^^^^ Use a string value for `class_name`.
+      RUBY
+
+      expect_correction(<<~RUBY)
+        has_many :accounts, class_name: "Account"
       RUBY
     end
 
@@ -25,12 +33,20 @@ RSpec.describe RuboCop::Cop::Rails::ReflectionClassName do
         has_one :account, class_name: Account
                           ^^^^^^^^^^^^^^^^^^^ Use a string value for `class_name`.
       RUBY
+
+      expect_correction(<<~RUBY)
+        has_one :account, class_name: "Account"
+      RUBY
     end
 
     it 'belongs_to' do
       expect_offense(<<~RUBY)
         belongs_to :account, class_name: Account
                              ^^^^^^^^^^^^^^^^^^^ Use a string value for `class_name`.
+      RUBY
+
+      expect_correction(<<~RUBY)
+        belongs_to :account, class_name: "Account"
       RUBY
     end
   end